BLAN-967 - all files of icons and pictures are renamed to one common format
